+prism.getPath()
getPath() will look at the corresponding meta object for each component and build a path from there.

This reduces path building redundany and makes the wrapper less fragile as the urls will automatically change if the prism project switches to a new direectory strucutre (assuming the meta component part is preserved of course).